[SD1-A07] Add xml supply
[git-p4: depot-paths = "//depot/C35/Students/shangliz/SD/Engine/": change = 442640]

+++ Code/Engine/Math/Vec2.cpp
@@ -326,6 +326,6 @@ void Vec2::SetFromText(char const* text)
         return;
     }
 
-    x = atof(parts[0].c_str());
-    y = atof(parts[1].c_str());
+    x = (float)atof(parts[0].c_str());
+    y = (float)atof(parts[1].c_str());
 }
